LP/Practica/requirements.txt
antlr4-python3-runtime==4.7.2
networkx==2.4
matplotlib==3.1.2
python-telegram-bot==12.2.0
